+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 - Fix: TACAN/ICLS (Also now visible in Kneeboard)
 - Fix: Waypoint Number in Kneeboard with multiple clients
 - Culling: For performance reasons we don't include ground groups and structures which are not a target or nearby a target. SHORAD/AA units are still in the mission!
-
+- Improved persistance modal
 
 ## 0.4.0 (2023-10-15)